pagination section redesign
