--- src/components/PagePagination copy/index.js
@@ -1,1 +1,0 @@
-export { default as default } from './PagePagination';
