+++ src/components/PagePagination/PagePagination.jsx
@@ -1,41 +1,18 @@
-import styles from './PagePagination.module.css';
+import React from 'react';
 import PropTypes from 'prop-types';
-import clsx from 'clsx';
-import Svg from '../Svg';
-import Link from 'next/link';
+import PagePaginationLink from '../PagePaginationLink';
 
-export default function PagePagination({ link, navDirection, text, ...props }) {
-  const pagePagination = clsx(styles.pagePagination, {
-    [styles.prevButton]: navDirection === 'previous',
-  });
+import styles from './PagePagination.module.css';
 
-  const direction = clsx(styles.direction, {
-    [styles.prevDirection]: navDirection === 'previous',
-  });
-
+export default function PagePagination({ text, link }) {
   return (
-    <Link href={link} passHref>
-      <a {...props} className={pagePagination} {...props}>
-        <div className={direction}>
-          <span>{navDirection}</span>
-          <Svg
-            height="18"
-            width="18"
-            href={`/icons/arrow-right-rounded-green-200.svg`}
-            title={`${navDirection} icon`}
-          />
-        </div>
-        <span>{`Go to "${text}"`}</span>
-      </a>
-    </Link>
+    <div className={styles.pagePagination}>
+      <PagePaginationLink text={text} link={link} />
+    </div>
   );
 }
-PagePagination.propTypes = {
-  link: PropTypes.string.isRequired,
-  navDirection: PropTypes.oneOf(['previous', 'next']),
-  text: PropTypes.string.isRequired,
-};
 
-PagePagination.defaultProps = {
-  navDirection: 'next',
+PagePagination.propTypes = {
+  text: PropTypes.string.isRequired,
+  link: PropTypes.string.isRequired,
 };
