+++ src/components/PagePaginationLink/index.js
@@ -1,0 +1,1 @@
+export { default as default } from './PagePaginationLink';
